move import source for reanalyses
