+++ globsim/era_interim.py
@@ -51,9 +51,12 @@ from os import path, listdir, remove
 from scipy.interpolate import interp1d
 from fnmatch import filter
 
+from globsim.GenericDownload import GenericDownload
+from globsim.GenericInterpolate import GenericInterpolate
+from globsim.GenericScale import GenericScale
+
 from globsim.common_utils import (series_interpolate, variables_skip, str_encode,
-                                  cummulative2total, GenericDownload, GenericScale,
-                                  GenericInterpolate)
+                                  cummulative2total)
 from globsim.nc_elements import netcdf_base, new_interpolated_netcdf, new_scaled_netcdf
 from globsim.meteorology import spec_hum_kgkg, LW_downward, pressure_from_elevation
